requirements.txt
# Agent orchestration
crewai==0.95.0

# RSS parsing
feedparser==6.0.11

# HTTP clients
httpx==0.27.0
requests==2.31.0

# AI APIs (OpenAI SDK used for OpenAI + Perplexity)
openai==1.51.0

# Telegram
python-telegram-bot==21.5

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Web / Admin panel
fastapi==0.115.0
sqladmin==0.19.0
uvicorn==0.30.6
jinja2==3.1.4

# Scheduler
apscheduler==3.10.4

# Retry logic
tenacity==8.5.0

# Validation
pydantic==2.9.2

# Env management
python-dotenv==1.0.1

# Security
passlib[bcrypt]==1.7.4

# Embeddings / dedup
# numpy>=1.26.4 требует Python 3.9+; Docker (3.11) получит 1.26.4
# Python 3.8 локально получает совместимую версию
numpy>=1.24.4

# Observability (optional)
langsmith==0.1.120
